pyproject.toml

[tool.poetry]
name = "aioworxmower"
version = "0.2.0"
description = "Async Python library for Worx Landroid mowers (OAuth, discovery, AWS IoT MQTT/WS, schedules)"
authors = ["aioworxmower Team <dev@example.com>"]
readme = "README.md"
packages = [{ include = "aioworxmower", from = "src" }]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
httpx = ">=0.27"
anyio = ">=4.0"
PyYAML = ">=6.0"
awscrt = "*"
awsiot = "*"
orjson = "^3.9.5"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
pytest-cov = ">=5.0"
ruff = ">=0.6"
pre-commit = ">=3.7"
mypy = ">=1.10"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
  "integration: tests that hit the real cloud (require secrets.yaml or GitHub Secrets)",
  "unit: offline tests with fixtures only"
]
